refactor: 초기 화면 랜더링 함수 분리
