+++ src/render/index.ts
@@ -1,9 +1,12 @@
-import registerRestaurantModal from './registerRestaurantModal';
+import init from './init';
 import restaurantList from './restaurantList';
+import registerRestaurantModal from './registerRestaurantModal';
 
 export default {
-  openRegisterRestaurantModal: registerRestaurantModal.open,
-  closeRegisterRestaurantModal: registerRestaurantModal.close,
+  init,
 
   restaurantList,
+
+  openRegisterRestaurantModal: registerRestaurantModal.open,
+  closeRegisterRestaurantModal: registerRestaurantModal.close,
 };
